requirements.txt
httpx
python-dotenv
flask